Remove dataclass requirement for python 3.6 and cleanup code
