+++ .gitignore
@@ -105,6 +105,7 @@ celerybeat.pid
 .env
 .venv
 env/
+env-*/
 venv/
 ENV/
 env.bak/
