+++ README.md
@@ -18,7 +18,6 @@ Requirements:
 
 * Python >= 3.5
 * FFMpeg >= 1.2 with the ffprobe command
-* dataclasses (only on exactly python 3.6.*)
 * Matplotlib python 3 library (install: `python3 -m pip install -U --user matplotlib`)
 
 Installation
