+++ requirements-dev.txt
@@ -1,3 +1,6 @@
 matplotlib
 pyqt5
-dataclasses;python_version=="3.6"
+setuptools
+wheel
+mypy
+pylint
