+++ requirements.txt
@@ -1,4 +1,3 @@
 matplotlib>=3.2.1;python_version>="3.6"
 matplotlib>=3.0.3;python_version=="3.5"
-dataclasses==0.7;python_version=="3.6"
 PyQt5>=5.14.2
